feat: add short IDs (chartNumber, shortId) for human-friendly patient and encounter URLs
- Add backend/api/resolvers.py with dual-accept resolver pattern (UUID or short ID)
- Patient URLs now accept chart_number (auto-increment integer starting 10001)
- Encounter URLs now accept short_id (8-char nanoid via secrets.token_urlsafe)
- Update all 12+ backend route files to resolve short IDs transparently
- Extend schemas, frontend types, stores, and components with new fields
- Update schedule, dashboard, patients pages to use short IDs in navigation
- Add SQL scripts for staff_user_id, tenant_slug, audit_log setup
- Add bootstrap/provisioning utilities for dev workflow

+++ backend/schemas/billing.py
@@ -104,7 +104,7 @@ class LineItemResponse(AppBaseModel):
     superbill_id: uuid.UUID
     cpt_code: str
     description: str
-    fee: Decimal
+    fee: float
     units: int
     diagnosis_pointers: list[str]
     modifiers: list[str]
@@ -178,7 +178,7 @@ class SuperbillResponse(AppBaseModel):
     mdm_level: str | None = None
     mdm_reasoning: str | None = None
     suggested_em_code: str | None = None
-    total_fee: Decimal
+    total_fee: float
     notes: str | None = None
     created_by_id: uuid.UUID | None = None
     line_items: list[LineItemResponse] = Field(default_factory=list)
